misp/requirements.txt
pymisp
urllib3
python-dateutil